HPのfloat化に伴う引数の型変更に対応

+++ entity/EntityAniBot.java
@@ -1,9 +1,6 @@
 package riseautomatons.entity;
 
-import net.minecraft.entity.EntityCreature;
-import net.minecraft.entity.monster.EntityGolem;
 import net.minecraft.entity.monster.EntityMob;
-import net.minecraft.entity.passive.EntityAnimal;
 import net.minecraft.util.DamageSource;
 import net.minecraft.world.World;
 
@@ -14,7 +11,7 @@ public abstract class EntityAniBot extends EntityMob {
 	}
 
 	@Override
-	public boolean attackEntityFrom(DamageSource damagesource, int i) {
+	public boolean attackEntityFrom(DamageSource damagesource, float i) {
 		if (damagesource == DamageSource.inWall) {
 			this.pushOutOfBlocks(posX, posY, posZ);
 			return false;
